feat: cache `dirs::data_dir` and `dirs::cache_dir`
